Reorganize components and clean up obsolete files
- Delete obsolete InfectionSequence.tsx
- Rename WelcomeGameScreen to DownloadScreen for clarity
- Move SystemCompromisedAnimation to login/ folder for future use
- Move InfectionProgressBar and HomeScreen to new home/ folder
- Remove empty infection/ folder
- Update all import paths and remove unused bugType props
- Clean up BUG_ASSIGNMENTS constant
