+++ app/components/home/InfectionProgressBar.tsx
