--- app/utils/infectionSequence.ts
@@ -1,76 +1,0 @@
-export interface InfectionStep {
-  glitchLevel: number;
-  terminalText: string;
-  delay: number;
-  action?: 'tower-defense' | 'emergency-mode';
-}
-
-export const INFECTION_STEPS: InfectionStep[] = [
-  {
-    glitchLevel: 1,
-    terminalText: '',
-    delay: 0
-  },
-  {
-    glitchLevel: 2,
-    terminalText: 'Booting in emergency mode...',
-    delay: 500
-  },
-  {
-    glitchLevel: 3,
-    terminalText: 'Booting in emergency mode...\nDoing a self-analysis...',
-    delay: 1000
-  },
-  {
-    glitchLevel: 4,
-    terminalText: 'Booting in emergency mode...\nDoing a self-analysis...\nSystem Status: INFECTED',
-    delay: 1500
-  },
-  {
-    glitchLevel: 5,
-    terminalText: 'Booting in emergency mode...\n[SYSTEM] Self-analyzing...\n[SYSTEM] Status: INFECTED\nSeverity: CRITICAL',
-    delay: 2000
-  },
-  {
-    glitchLevel: 6,
-    terminalText: 'Booting in emergency mode...\nDoing a self-analysis...\nSystem Status: INFECTED\nSeverity: CRITICAL\nSource: fake DOWNLOAD button, device infected with Spartan virus',
-    delay: 2500
-  },
-  {
-    glitchLevel: 0,
-    terminalText: '',
-    delay: 3000,
-    action: 'tower-defense'
-  }
-];
-
-export const TOTAL_INFECTION_TIME = 3000;
-
-export function startInfectionSequence(
-  setGlitchLevel: (level: number) => void,
-  setTerminalText: (text: string) => void,
-  onComplete: () => void
-) {
-  // Start infection
-  setGlitchLevel(1);
-  
-  // Schedule each step
-  INFECTION_STEPS.forEach((step, index) => {
-    if (index === 0) return; // Skip the first step (already set)
-    
-    setTimeout(() => {
-      setGlitchLevel(step.glitchLevel);
-      setTerminalText(step.terminalText);
-      
-      // If this is the last step, schedule completion
-      if (index === INFECTION_STEPS.length - 1) {
-        setTimeout(() => {
-          onComplete();
-        }, TOTAL_INFECTION_TIME - step.delay);
-      }
-    }, step.delay);
-  });
-}
-
-// Default export to satisfy Expo Router
-export default INFECTION_STEPS;
